+++ src/cli/command.rs
@@ -36,7 +36,7 @@ use crate::bus::{
     info::{Address, InfoMsg},
     AddressSecretKey,
 };
-use crate::bus::{BusMsg, Failure, FailureCode};
+use crate::bus::{BusMsg, Failure, FailureCode, HealthReport};
 use crate::client::Client;
 use crate::syncerd::{SweepAddressAddendum, SweepBitcoinAddress, SweepMoneroAddress};
 use crate::{Error, LogStyle, ServiceId};
@@ -176,6 +176,96 @@ impl Exec for Command {
             Command::Connect { swap_id } => {
                 runtime.request_ctl(ServiceId::Farcasterd, CtlMsg::Connect(swap_id))?;
                 runtime.report_response_or_fail()?;
+            }
+
+            Command::HealthCheck => {
+                runtime.request_ctl(
+                    ServiceId::Farcasterd,
+                    CtlMsg::HealthCheck(Blockchain::Bitcoin, Network::Testnet),
+                )?;
+                let bitcoin_testnet_health = match runtime.response()? {
+                    BusMsg::Ctl(CtlMsg::HealthResult(health)) => health,
+                    _ => {
+                        return Err(Error::Other(
+                            "Server returned unexpected response for call health check".to_string(),
+                        ))
+                    }
+                };
+
+                runtime.request_ctl(
+                    ServiceId::Farcasterd,
+                    CtlMsg::HealthCheck(Blockchain::Bitcoin, Network::Mainnet),
+                )?;
+                let bitcoin_mainnet_health = match runtime.response()? {
+                    BusMsg::Ctl(CtlMsg::HealthResult(health)) => health,
+                    _ => {
+                        return Err(Error::Other(
+                            "Server returned unexpected response for call health check".to_string(),
+                        ))
+                    }
+                };
+                runtime.request_ctl(
+                    ServiceId::Farcasterd,
+                    CtlMsg::HealthCheck(Blockchain::Bitcoin, Network::Local),
+                )?;
+                let bitcoin_local_health = match runtime.response()? {
+                    BusMsg::Ctl(CtlMsg::HealthResult(health)) => health,
+                    _ => {
+                        return Err(Error::Other(
+                            "Server returned unexpected response for call health check".to_string(),
+                        ))
+                    }
+                };
+
+                runtime.request_ctl(
+                    ServiceId::Farcasterd,
+                    CtlMsg::HealthCheck(Blockchain::Monero, Network::Testnet),
+                )?;
+                let monero_testnet_health = match runtime.response()? {
+                    BusMsg::Ctl(CtlMsg::HealthResult(health)) => health,
+                    _ => {
+                        return Err(Error::Other(
+                            "Server returned unexpected response for call health check".to_string(),
+                        ))
+                    }
+                };
+
+                runtime.request_ctl(
+                    ServiceId::Farcasterd,
+                    CtlMsg::HealthCheck(Blockchain::Monero, Network::Mainnet),
+                )?;
+                let monero_mainnet_health = match runtime.response()? {
+                    BusMsg::Ctl(CtlMsg::HealthResult(health)) => health,
+                    _ => {
+                        return Err(Error::Other(
+                            "Server returned unexpected response for call health check".to_string(),
+                        ))
+                    }
+                };
+
+                runtime.request_ctl(
+                    ServiceId::Farcasterd,
+                    CtlMsg::HealthCheck(Blockchain::Monero, Network::Local),
+                )?;
+                let monero_local_health = match runtime.response()? {
+                    BusMsg::Ctl(CtlMsg::HealthResult(health)) => health,
+                    _ => {
+                        return Err(Error::Other(
+                            "Server returned unexpected response for call health check".to_string(),
+                        ))
+                    }
+                };
+
+                let report = HealthReport {
+                    bitcoin_testnet_health,
+                    bitcoin_mainnet_health,
+                    bitcoin_local_health,
+                    monero_testnet_health,
+                    monero_mainnet_health,
+                    monero_local_health,
+                };
+
+                println!("{}", report);
             }
 
             Command::Make {
@@ -403,7 +493,7 @@ impl Exec for Command {
                     InfoMsg::GetAddressSecretKey(Address::Bitcoin(source_address.clone())),
                 )?;
                 if let BusMsg::Info(InfoMsg::AddressSecretKey(AddressSecretKey::Bitcoin {
-                    secret_key,
+                    secret_key_info,
                     ..
                 })) = runtime.report_failure()?
                 {
@@ -411,7 +501,7 @@ impl Exec for Command {
                         ServiceId::Farcasterd,
                         CtlMsg::SweepAddress(SweepAddressAddendum::Bitcoin(SweepBitcoinAddress {
                             source_address,
-                            source_secret_key: secret_key,
+                            source_secret_key: secret_key_info.secret_key,
                             destination_address,
                         })),
                     )?;
@@ -430,16 +520,15 @@ impl Exec for Command {
                     InfoMsg::GetAddressSecretKey(Address::Monero(source_address)),
                 )?;
                 if let BusMsg::Info(InfoMsg::AddressSecretKey(AddressSecretKey::Monero {
-                    view,
-                    spend,
+                    secret_key_info,
                     ..
                 })) = runtime.report_failure()?
                 {
                     runtime.request_ctl(
                         ServiceId::Farcasterd,
                         CtlMsg::SweepAddress(SweepAddressAddendum::Monero(SweepMoneroAddress {
-                            source_spend_key: spend,
-                            source_view_key: view,
+                            source_spend_key: secret_key_info.spend,
+                            source_view_key: secret_key_info.view,
                             destination_address,
                             minimum_balance: monero::Amount::from_pico(0),
                         })),